+++ CHANGELOG.md
@@ -6,13 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 #### 2025-08-31
+
 **Added:**
+- **Component updates in VirtualKeyboard**
+- **Component updates in +page**
+
 - **Component updates in VirtualKeyboard**
 
 **Changed:**
-- **Documentation**: Updated project documentation
 
+- **Documentation**: Updated project documentation
 
 #### 2025-08-30